% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrdif_total.R
\name{lrdif_total}
\alias{lrdif_total}
\title{Using logistic Regression Method to Detect DIF for
all the items in the dataset}
\usage{
lrdif_total(filename)
}
\arguments{
\item{filename}{filename is the name of the dataset you want to test for DIF.}
}
\description{
This function use logistic regression method to detect DIF for
all the items in the dataset.
}
\details{
The logistic regression is a commonly used method to
dectect DIF.
}

