% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test1.R
\docType{data}
\name{test1}
\alias{test1}
\title{This is data to be included in lrdif package}
\format{A 2000 by 16 \code{data.frame} The first column i1 which corresponds to group membership
(The Group is coded 0=reference, 1=focal). i2 i3... i15 are dichotomous ability items.}
\description{
This dataset provided by Dr. Anna Brown in the course of Measurement invariance and
Differential Item Functioning.
}
\references{
\url{http://www.restore.ac.uk/appliedpsychometrics/Tutorial.materials/Measurement.invariance/}
}
\keyword{data}

