% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrdif.R
\name{lrdif}
\alias{lrdif}
\title{Using logistic Regression Method to Detect DIF}
\usage{
lrdif(Data, itemn, groupcolumn = 1, alpha = 0.05)
}
\arguments{
\item{Data}{matrix or data.frame of scored item responses.}
}
\description{
This function use logistic regression to detect
dichotomous differential item functioning.
}
\details{
The logistic regression is a commonly used method to
dectect DIF.
}

